Addressing many of warnings emitted w/ -pedantic -Wall -Wextra (#643)
* warning: unused parameter ‘sig’ [-Wunused-parameter]

* warning: comparison of integer expressions of different signedness

* warning: ‘total_time’ may be used uninitialized

* warning: unused variable ‘cl_cpu_device’ + warning: unused variable ‘cl_gpu_device’

* comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Wsign-compare]

* warning: comparison of unsigned expression in ‘>= 0’ is always true

* warning: comparison of integer expressions of different signedness: ‘int’ and ‘uint32_t’ {aka ‘unsigned int’} [-Wsign-compare]

* warning: comparison of integer expressions of different signedness: ‘int’ and ‘std::vector<std::__cxx11::basic_string<char> >::size_type’ {aka ‘long unsigned int’} [-Wsign-compare]

* warning: unused parameter ‘xxxx’ [-Wunused-parameter]

* warning: unused parameter ‘xxxx’ [-Wunused-parameter]

* warning: unused parameter ‘xxxx’ [-Wunused-parameter] + use of for range loops

* warning: initializator in constructor reordered

* Fix unused parameters

* Fix signedness and constructor order

* warning: missing initializer for member ‘timespec::tv_nsec’ [-Wmissing-field-initializers]

* Fix signedness

* Unused parameter

+++ tools/unitrace/src/chromelogger.h
@@ -398,7 +398,7 @@ static std::string convertDataToString(IttArgs* args) {
     switch (args->type) {
       case __itt_metadata_u64: {
         const uint64_t* uint64Ptr = reinterpret_cast<const uint64_t*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
@@ -408,7 +408,7 @@ static std::string convertDataToString(IttArgs* args) {
       }
       case __itt_metadata_s64: {
         const int64_t* int64Ptr = reinterpret_cast<const int64_t*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
@@ -418,7 +418,7 @@ static std::string convertDataToString(IttArgs* args) {
       }
       case __itt_metadata_u32: {
         const uint32_t* uint32Ptr = reinterpret_cast<const uint32_t*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
@@ -428,7 +428,7 @@ static std::string convertDataToString(IttArgs* args) {
       }
       case __itt_metadata_s32: {
         const int32_t* int32Ptr = reinterpret_cast<const int32_t*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
@@ -438,7 +438,7 @@ static std::string convertDataToString(IttArgs* args) {
       }
       case __itt_metadata_u16: {
         const uint16_t* uint16Ptr = reinterpret_cast<const uint16_t*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
@@ -448,7 +448,7 @@ static std::string convertDataToString(IttArgs* args) {
       }
       case __itt_metadata_s16: {
         const int16_t* uint16Ptr = reinterpret_cast<const int16_t*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
@@ -458,7 +458,7 @@ static std::string convertDataToString(IttArgs* args) {
       }
       case __itt_metadata_float: {
         const float* floatPtr = reinterpret_cast<const float*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
@@ -468,7 +468,7 @@ static std::string convertDataToString(IttArgs* args) {
       }
       case __itt_metadata_double: {
         const double* doublePtr = reinterpret_cast<const double*>(dataPtr);
-        for (int i=0; i < args->count; i++) {
+        for (size_t i=0; i < args->count; i++) {
           if (i) {
             strData += ",";
           }
